LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-constructor.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSFullscreenMediaControls</code> constructor.");

const mediaControls = new MacOSFullscreenMediaControls;

shouldBeEqualToString("mediaControls.element.localName", "div");
shouldBeEqualToString("mediaControls.element.className", "media-controls mac fullscreen");
shouldBe("mediaControls.layoutTraits", "LayoutTraits.macOS | LayoutTraits.Fullscreen");

debug("");
debug("Controls bar");
shouldBe("mediaControls.controlsBar.children.length", "4");
shouldBe("mediaControls.controlsBar.children[0]", "mediaControls.volumeSlider");
shouldBeTrue("mediaControls.controlsBar.children[1] instanceof ButtonsContainer");
shouldBeTrue("mediaControls.controlsBar.children[2] instanceof ButtonsContainer");
shouldBe("mediaControls.controlsBar.children[3]", "mediaControls.timeControl");

debug("");
debug("Center container");
shouldBe("mediaControls.controlsBar.children[1].buttons", "[mediaControls.rewindButton, mediaControls.playPauseButton, mediaControls.forwardButton]");
shouldBeEqualToString("mediaControls.controlsBar.children[1].element.className", "buttons-container center");
shouldBe("mediaControls.controlsBar.children[1].padding", "27");
shouldBe("mediaControls.controlsBar.children[1].margin", "27");

debug("");
debug("Right container");
shouldBe("mediaControls.controlsBar.children[2].buttons", "[mediaControls.airplayButton, mediaControls.pipButton, mediaControls.tracksButton, mediaControls.fullscreenButton]");
shouldBeEqualToString("mediaControls.controlsBar.children[2].element.className", "buttons-container right");
shouldBe("mediaControls.controlsBar.children[2].padding", "12");
shouldBe("mediaControls.controlsBar.children[2].margin", "0");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
