LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-constructor-expected.txt

Testing the MacOSFullscreenMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "media-controls mac fullscreen"
PASS mediaControls.layoutTraits is LayoutTraits.macOS | LayoutTraits.Fullscreen

Controls bar
PASS mediaControls.controlsBar.children.length is 4
PASS mediaControls.controlsBar.children[0] is mediaControls.volumeSlider
PASS mediaControls.controlsBar.children[1] instanceof ButtonsContainer is true
PASS mediaControls.controlsBar.children[2] instanceof ButtonsContainer is true
PASS mediaControls.controlsBar.children[3] is mediaControls.timeControl

Center container
PASS mediaControls.controlsBar.children[1].buttons is [mediaControls.rewindButton, mediaControls.playPauseButton, mediaControls.forwardButton]
PASS mediaControls.controlsBar.children[1].element.className is "buttons-container center"
PASS mediaControls.controlsBar.children[1].padding is 27
PASS mediaControls.controlsBar.children[1].margin is 27

Right container
PASS mediaControls.controlsBar.children[2].buttons is [mediaControls.airplayButton, mediaControls.pipButton, mediaControls.tracksButton, mediaControls.fullscreenButton]
PASS mediaControls.controlsBar.children[2].element.className is "buttons-container right"
PASS mediaControls.controlsBar.children[2].padding is 12
PASS mediaControls.controlsBar.children[2].margin is 0

PASS successfullyParsed is true

TEST COMPLETE
